Separated Js into modules
location & pollen in separate files
